Organized hooks into sub-module and create Histotical negative sampler
