+++ test/unit/test_hooks/test_negative_edge_sampler_hook.py
@@ -1,12 +1,14 @@
-from unittest.mock import patch
-
 import pytest
 import torch
 
 from tgm import DGBatch, DGraph
 from tgm.constants import PADDED_NODE_ID
 from tgm.data import DGData, DGDataLoader
-from tgm.hooks import HookManager, NegativeEdgeSamplerHook
+from tgm.hooks import (
+    HistoricalNegativeEdgeSamplerHook,
+    HookManager,
+    RandomNegativeEdgeSamplerHook,
+)
 
 
 @pytest.fixture
@@ -17,36 +19,48 @@ def data():
 
 
 def test_hook_dependancies():
-    hook = NegativeEdgeSamplerHook(low=0, high=10)
+    hook = RandomNegativeEdgeSamplerHook(low=0, high=10)
     assert hook.requires == {'edge_src', 'edge_dst', 'edge_time'}
     assert hook.produces == {'neg', 'neg_time'}
 
-    hook_with_id = NegativeEdgeSamplerHook(low=0, high=10, id='foo')
+    hook_with_id = RandomNegativeEdgeSamplerHook(low=0, high=10, id='foo')
     assert hook_with_id.requires == {'edge_src', 'edge_dst', 'edge_time'}
     assert hook_with_id.produces == {'neg_foo', 'neg_time_foo'}
 
+    hook = HistoricalNegativeEdgeSamplerHook()
+    assert hook.requires == {'edge_src', 'edge_dst', 'edge_time'}
+    assert hook.produces == {'neg', 'neg_time', 'valid_neg_mask'}
+
+    hook_with_id = HistoricalNegativeEdgeSamplerHook(id='foo')
+    assert hook_with_id.requires == {'edge_src', 'edge_dst', 'edge_time'}
+    assert hook_with_id.produces == {'neg_foo', 'neg_time_foo', 'valid_neg_mask_foo'}
+
 
 def test_hook_repre():
-    hook_with_id = NegativeEdgeSamplerHook(low=0, high=10, id='foo')
+    hook_with_id = RandomNegativeEdgeSamplerHook(low=0, high=10, id='foo')
+    assert 'foo' in hook_with_id.__repr__()
+
+    hook_with_id = HistoricalNegativeEdgeSamplerHook(id='foo')
     assert 'foo' in hook_with_id.__repr__()
 
 
 def test_hook_reset_state():
-    assert NegativeEdgeSamplerHook.has_state == True
+    assert RandomNegativeEdgeSamplerHook.has_state == False
+    assert HistoricalNegativeEdgeSamplerHook.has_state == True
 
 
 def test_bad_negative_edge_sampler_init():
     with pytest.raises(ValueError):
-        NegativeEdgeSamplerHook(low=0, high=0)
+        RandomNegativeEdgeSamplerHook(low=0, high=0)
     with pytest.raises(ValueError):
-        NegativeEdgeSamplerHook(low=0, high=1, neg_ratio=0)
+        RandomNegativeEdgeSamplerHook(low=0, high=1, neg_ratio=0)
     with pytest.raises(ValueError):
-        NegativeEdgeSamplerHook(low=0, high=1, neg_ratio=2)
+        RandomNegativeEdgeSamplerHook(low=0, high=1, neg_ratio=2)
 
 
 def test_negative_edge_sampler(data):
     dg = DGraph(data)
-    hook = NegativeEdgeSamplerHook(low=0, high=10)
+    hook = RandomNegativeEdgeSamplerHook(low=0, high=10)
     batch = hook(dg, dg.materialize())
     assert isinstance(batch, DGBatch)
     assert torch.is_tensor(batch.neg)
@@ -57,7 +71,7 @@ def test_negative_edge_sampler(data):
 
 def test_negative_edge_sampler_with_id(data):
     dg = DGraph(data)
-    hook = NegativeEdgeSamplerHook(low=0, high=10, id='foo')
+    hook = RandomNegativeEdgeSamplerHook(low=0, high=10, id='foo')
     batch = hook(dg, dg.materialize())
     assert isinstance(batch, DGBatch)
     assert torch.is_tensor(batch.neg_foo)
@@ -85,7 +99,7 @@ def node_only_data():
 def test_node_only_batch_negative_edge_sampler(node_only_data):
     dg = DGraph(node_only_data)
     hm = HookManager(keys=['unit'])
-    hm.register('unit', NegativeEdgeSamplerHook(low=0, high=6))
+    hm.register('unit', RandomNegativeEdgeSamplerHook(low=0, high=6))
     loader = DGDataLoader(dg, batch_size=3, hook_manager=hm)
     with hm.activate('unit'):
         batch_iter = iter(loader)
@@ -103,81 +117,84 @@ def test_node_only_batch_negative_edge_sampler(node_only_data):
 
 
 @pytest.fixture
-def data_test_hst_rnd():
+def data_test_hst_sampling():
     edge_index = torch.IntTensor(
         [
+            # 1st batch
             [1, 5],
             [7, 6],
             [2, 8],
-            [7, 8],  # 1st batch
+            [7, 8],
+            # 2nd batch
             [1, 7],
             [9, 10],
             [3, 10],
-            [1, 9],  # 2nd batch
+            [1, 9],
+            # 3rd batch
             [3, 11],
             [2, 10],
             [7, 2],
-            [3, 5],  # 3rd batch
+            [3, 5],
         ]
     )
     edge_time = torch.arange(edge_index.size(0))
     return DGData.from_raw(edge_time, edge_index)
 
 
-def test_hst_rnd(data_test_hst_rnd):
-    dg = DGraph(data_test_hst_rnd)
+def test_hst_sampling(data_test_hst_sampling):
+    dg = DGraph(data_test_hst_sampling)
 
     hm = HookManager(keys=['unit'])
-    sampler = NegativeEdgeSamplerHook(low=0, high=6, strategy='hist_rnd')
-
-    def mock_random_sampling(dg: DGraph, batch: DGBatch):
-        neg = torch.full((batch.edge_src.size(0),), PADDED_NODE_ID, dtype=torch.int32)
-        neg_time = batch.edge_time.clone()
-        return neg, neg_time
+    sampler = HistoricalNegativeEdgeSamplerHook()
 
     hm.register('unit', sampler)
     loader = DGDataLoader(dg, batch_size=4, hook_manager=hm)
 
-    with patch.object(sampler, '_random_sampling', mock_random_sampling):
-        with hm.activate('unit'):
-            batch_iter = iter(loader)
-            batch_1 = next(batch_iter)
-            assert batch_1.neg.shape == (4,)
-            assert torch.equal(
-                batch_1.neg,
-                torch.Tensor(
-                    [PADDED_NODE_ID, PADDED_NODE_ID, PADDED_NODE_ID, PADDED_NODE_ID]
-                ),
-            )
-            assert sampler._memory is not None
-            assert sampler._memory.shape == (2, 8)
-            assert sampler._count == 4
+    with hm.activate('unit'):
+        batch_iter = iter(loader)
+        batch_1 = next(batch_iter)
+        assert batch_1.neg.shape == (4,)
+        assert torch.equal(
+            batch_1.neg,
+            torch.Tensor(
+                [PADDED_NODE_ID, PADDED_NODE_ID, PADDED_NODE_ID, PADDED_NODE_ID]
+            ),
+        )
+        assert torch.equal(
+            batch_1.valid_neg_mask,
+            torch.Tensor([False, False, False, False]),
+        )
+        assert sampler._memory is not None
+        assert sampler._memory.shape == (2, 8)
+        assert sampler._count == 4
 
-            batch_2 = next(batch_iter)
-            assert batch_2.neg.shape == (4,)
-            assert torch.equal(
-                batch_2.neg, torch.Tensor([5, PADDED_NODE_ID, PADDED_NODE_ID, 5])
-            )
-            assert sampler._memory is not None
-            assert sampler._memory.shape == (2, 8)
-            assert sampler._count == 8
+        batch_2 = next(batch_iter)
+        assert batch_2.neg.shape == (4,)
+        assert torch.equal(
+            batch_2.neg, torch.Tensor([5, PADDED_NODE_ID, PADDED_NODE_ID, 5])
+        )
+        assert torch.equal(
+            batch_2.valid_neg_mask,
+            torch.Tensor([True, False, False, True]),
+        )
+        assert sampler._memory is not None
+        assert sampler._memory.shape == (2, 8)
+        assert sampler._count == 8
 
-            batch_3 = next(batch_iter)
-            assert batch_3.neg.shape == (4,)
-            assert torch.equal(
-                batch_3.neg, torch.Tensor([10, 8, PADDED_NODE_ID, PADDED_NODE_ID])
-            )
+        batch_3 = next(batch_iter)
+        assert batch_3.neg.shape == (4,)
+        assert torch.equal(batch_3.neg, torch.Tensor([10, 8, 8, 10])) or torch.equal(
+            batch_3.neg, torch.Tensor([10, 8, 6, 10])
+        )
+        assert torch.equal(
+            batch_3.valid_neg_mask,
+            torch.Tensor([True, True, True, True]),
+        )
+        assert sampler._memory is not None
+        assert sampler._memory.shape == (2, 24)
+        assert sampler._count != 0
+        assert sampler._count == 12
 
-            assert sampler._memory is not None
-            assert sampler._memory.shape == (2, 24)
-            assert sampler._count != 0
-            assert sampler._count == 12
-
-            sampler.reset_state()
-            assert sampler._memory is None
-            assert sampler._count == 0
-
-
-def test_invalid_strategy():
-    with pytest.raises(ValueError):
-        NegativeEdgeSamplerHook(low=0, high=6, strategy='foo')
+        sampler.reset_state()
+        assert sampler._memory is None
+        assert sampler._count == 0
