+++ tgm/hooks/__init__.py
@@ -2,7 +2,8 @@ from .base import DGHook, StatelessHook, StatefulHook, BaseDGHook, SeedableHook
 from .dedup import DeduplicationHook
 from .device import DeviceTransferHook, PinMemoryHook
 from .negatives import (
-    NegativeEdgeSamplerHook,
+    RandomNegativeEdgeSamplerHook,
+    HistoricalNegativeEdgeSamplerHook,
     TGBNegativeEdgeSamplerHook,
     TGBTHGNegativeEdgeSamplerHook,
     TGBTKGNegativeEdgeSamplerHook,
@@ -11,5 +12,4 @@ from .neighbors import NeighborSamplerHook, RecencyNeighborHook
 from .hook_manager import HookManager
 from .recipe import RecipeRegistry
 from .node_tracks import EdgeEventsSeenNodesTrackHook
-from .batch_analytics import BatchAnalyticsHook
-from .node_analytics import NodeAnalyticsHook
+from .analytics import BatchAnalyticsHook, NodeAnalyticsHook
